refactor: SignUpController, EmailController 통합
